% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all.R
\name{check_all}
\alias{check_all}
\title{Verifica todos os testes de validação}
\usage{
check_all(
  programas_planejamento,
  acoes_planejamento,
  localizadores_todos_planejamento,
  indicadores_planejamento,
  base_categoria_pessoal,
  base_detalhamento_obras,
  base_intra_orcamentaria_detalhamento,
  base_intra_orcamentaria_repasse,
  base_limite_cota,
  base_orcam_despesa_item_fiscal,
  base_orcam_receita_fiscal,
  base_orcam_receita_investimento,
  base_qdd_fiscal,
  base_qdd_investimento,
  base_qdd_plurianual,
  base_qdd_plurianual_invest,
  base_repasse_recursos,
  desc_setor_governo,
  uo_acao_inativo_civil,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verifica todos os testes de validação
}
