% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_area_tematica_exists.R
\name{check_area_tematica_exists}
\alias{check_area_tematica_exists}
\alias{check_area_tematica_exists_programas}
\alias{check_area_tematica_exists_acoes}
\alias{check_area_tematica_exists_localizadores}
\title{Área temática por programa}
\usage{
check_area_tematica_exists_programas(programas_planejamento)

check_area_tematica_exists_acoes(acoes_planejamento)

check_area_tematica_exists_localizadores(localizadores_todos_planejamento)
}
\description{
Verifica se cada programa possui apenas uma área temática. É obrigatório a existência de uma área por programa (ie. one and only one).

A verificação é realizada por base de dados:
\itemize{
\item programas_planejamento
\item acoes_planejamento
\item localizadores_todos_planejamento
}
}
