% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_area_tematica_exists_localizadores.R
\name{check_area_tematica_exists_localizadores}
\alias{check_area_tematica_exists_localizadores}
\title{Área temática por programa}
\usage{
check_area_tematica_exists_localizadores(
  localizadores_todos_planejamento,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Verifica se cada programa possui apenas uma área temática. É obrigatório a existência de uma área por programa (ie. \emph{one and only one}).
}
