% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_area_tematica_exists_programas.R
\name{check_area_tematica_exists_programas}
\alias{check_area_tematica_exists_programas}
\title{Verifica se cada programa possui apenas uma área temática - programas_planejamento}
\usage{
check_area_tematica_exists_programas(
  programas_planejamento,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
É obrigatório a existência de uma área por programa (ie. \emph{one and only one}).
}
