% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_composite_primary_key.R
\name{check_composite_primary_key_programas}
\alias{check_composite_primary_key_programas}
\title{Verifica chaves das bases}
\usage{
check_composite_primary_key_programas(
  programas_planejamento,
  stop_on_failure = FALSE,
  output = FALSE,
  json_outfile = NULL,
  log_level = "ERROR",
  msg_template = NULL
)
}
\description{
Existência de linhas duplicadas
}
