% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count.R
\name{check_count}
\alias{check_count}
\title{Verificar quantidade de programas e ações entre as bases (incluindo excluídos e incluídos)}
\description{
Verificar quantidade de programas e ações entre as bases (incluindo excluídos e incluídos)
}
