% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count_acoes.R
\name{check_count_acoes}
\alias{check_count_acoes}
\title{Verificar quantidade de ações entre as bases}
\usage{
check_count_acoes(
  acoes_planejamento,
  localizadores_todos_planejamento,
  output = FALSE,
  stop_on_failure = FALSE,
  json_outfile = NULL,
  log_level = "ERROR",
  msg_template = NULL
)
}
\description{
É esperado que haja a mesma quantidade de ações nas bases de ações e de localizadores,
ambas com a mesma descrição para cada ação.
}
