% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count_acoes_is_new.R
\name{check_count_acoes_is_new}
\alias{check_count_acoes_is_new}
\title{Verificar quantidade de programas e ações entre as bases (incluindo excluídos e incluídos)}
\usage{
check_count_acoes_is_new(
  acoes_planejamento,
  localizadores_todos_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verificar quantidade de programas e ações entre as bases (incluindo excluídos e incluídos)
}
