% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count_programas.R
\name{check_count_programas}
\alias{check_count_programas}
\title{Verificar quantidade de programas entre as bases}
\usage{
check_count_programas(
  programas_planejamento,
  acoes_planejamento,
  localizadores_todos_planejamento,
  output = FALSE,
  stop_on_failure = FALSE,
  json_outfile = NULL,
  log_level = "ERROR",
  msg_template = NULL
)
}
\description{
É esperado que haja a mesma quantidade de programas nas bases de programas,
ações e de localizadores, todas com a mesma descrição para cada programa.
}
