% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count_programas_is_new.R
\name{check_count_programas_is_new}
\alias{check_count_programas_is_new}
\title{Verificar quantidade de programas incluídos entre as bases}
\usage{
check_count_programas_is_new(
  programas_planejamento,
  acoes_planejamento,
  output = FALSE,
  stop_on_failure = FALSE,
  json_outfile = NULL,
  log_level = "ERROR",
  msg_template = NULL
)
}
\description{
Confere se o parâmetro is_new está idêntico entre as bases.
}
