% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_detalhamento_obras_acoes_exists.R
\name{check_detalhamento_obras_acoes_exists}
\alias{check_detalhamento_obras_acoes_exists}
\title{Verifica exclusão do detalhamento de obras de ações excluídas}
\usage{
check_detalhamento_obras_acoes_exists(
  base_detalhamento_obras,
  acoes_planejamento,
  stop_on_failure = FALSE,
  output = FALSE,
  json_outfile = NULL,
  log_level = "ERROR",
  msg_template = NULL
)
}
\description{
Espera-se que todas as ações registradas como deletadas não tenham
valores registrados na base de detalhamento de obras.
}
