% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_detalhamento_pessoal_inativo_civil.R
\name{check_detalhamento_pessoal_inativo_civil}
\alias{check_detalhamento_pessoal_inativo_civil}
\title{Verifica se o detalhamento do quantitativo de inativos civis foi realizado}
\usage{
check_detalhamento_pessoal_inativo_civil(
  base_categoria_pessoal,
  base_qdd_fiscal,
  uo_acao_inativo_civil,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
O detalhamento é obrigatório para ações orçamentárias que apresentam valores
superiores a R$ 1.000 nas ações do FFP.
}
