% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_equivalencia_sigplan_sisor.R
\name{check_equivalencia_sigplan_sisor}
\alias{check_equivalencia_sigplan_sisor}
\title{Total do orçamento fiscal e investimento SIGPLAN vs SISOR}
\description{
Verificar se o valor total do orçamento (orçamento fiscal e orçamento de
investimento das empresas controladas) coincide com a projeção do PPAG
para o ano seguinte
}
