% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fechamento_fonte_orcam_investimento.R
\name{check_fechamento_fonte_orcam_investimento}
\alias{check_fechamento_fonte_orcam_investimento}
\title{Fechamento receita e despesa por fonte do orçamento de investimento}
\usage{
check_fechamento_fonte_orcam_investimento(
  base_orcam_receita_investimento,
  base_qdd_investimento,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Fechamento receita e despesa por fonte do orçamento de investimento
}
