% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/check_indicadores_consistency_indice_de_referencia_exists.R
\name{check_indicadores_consistency_indice_de_referencia_exists}
\alias{check_indicadores_consistency_indice_de_referencia_exists}
\title{Verifica se existem indicadores com índice, mas sem metas previstas}
\usage{
check_indicadores_consistency_indice_de_referencia_exists(
  indicadores_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verifica se existem indicadores com índice, mas sem metas previstas
}
