% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/check_indicadores_consistency_indice_de_referencia_missing.R
\name{check_indicadores_consistency_indice_de_referencia_missing}
\alias{check_indicadores_consistency_indice_de_referencia_missing}
\title{Verifica se existem indicadores sem índice, mas com metas previstas}
\usage{
check_indicadores_consistency_indice_de_referencia_missing(
  indicadores_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verifica se existem indicadores sem índice, mas com metas previstas
}
