% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_indicadores_indice_referencia_exists.R
\name{check_indicadores_indice_referencia_exists}
\alias{check_indicadores_indice_referencia_exists}
\title{Verifica se indicadores apurados possuem índice de referência}
\usage{
check_indicadores_indice_referencia_exists(
  indicadores_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Se o indicador não está em apuração, ele não pode estar vazio.
}
