% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_indicadores_indice_referencia_zerado.R
\name{check_indicadores_indice_referencia_zerado}
\alias{check_indicadores_indice_referencia_zerado}
\title{Verifica se indicadores não possuem índice de referência zerados}
\usage{
check_indicadores_indice_referencia_zerado(
  indicadores_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verifica se indicadores não possuem índice de referência zerados
}
