% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/check_indicadores_justificativa_indice_referencia_em_apuracao.R
\name{check_indicadores_justificativa_indice_referencia_em_apuracao}
\alias{check_indicadores_justificativa_indice_referencia_em_apuracao}
\title{Verifica se indicadores com índice de referência “em apuração” possuem justificativa}
\usage{
check_indicadores_justificativa_indice_referencia_em_apuracao(
  indicadores_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Somente são aceitas como válidas justificativas acima de 50 caracteres
}
