% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_indicadores_previsoes_exists.R
\name{check_indicadores_previsoes_exists}
\alias{check_indicadores_previsoes_exists}
\title{Verifica se indicadores com previsões apuradas possuem previsões}
\usage{
check_indicadores_previsoes_exists(
  indicadores_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Se a previsão do indicador não está em apuração, ele não pode estar vazio.
}
