% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_indicadores_previsoes_zeradas.R
\name{check_indicadores_previsoes_zeradas}
\alias{check_indicadores_previsoes_zeradas}
\title{Verifica se indicadores n찾o possuem previs천es zeradas}
\usage{
check_indicadores_previsoes_zeradas(
  indicadores_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verifica se indicadores n찾o possuem previs천es zeradas
}
