% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_intra_despesa.R
\name{check_intra_despesa}
\alias{check_intra_despesa}
\title{Receitas e Despesas Intra-Orcamentárias}
\usage{
check_intra_despesa(
  base_orcam_despesa_item_fiscal,
  base_intra_orcamentaria_repasse,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Verifica se as operações intra-orçamentárias registradas na
base intraorçamentária de repasse estão adequadamente detalhadas no orçamento.
}
