% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_intra_detalhamento.R
\name{check_intra_detalhamento}
\alias{check_intra_detalhamento}
\title{Receitas e Despesas Intra-Orcamentárias}
\usage{
check_intra_detalhamento(
  base_intra_orcamentaria_detalhamento,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Verifica se as operações intra-orçamentárias registradas na
base intraorçamentária de repasse estão adequadamente detalhadas no orçamento.
}
