% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_limite.R
\name{check_limite}
\alias{check_limite}
\title{Valor Limite anual = Valor Utilizado + Valor Transferido}
\usage{
check_limite(base_limite_cota, stop_on_failure = FALSE, output = FALSE)
}
\description{
Valor Limite anual = Valor Utilizado + Valor Transferido
}
