% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_limite_ano0.R
\name{check_limite_ano0}
\alias{check_limite_ano0}
\title{Valor Limite anual = Valor Utilizado + Valor Transferido}
\usage{
check_limite_ano0(base_limite_cota, stop_on_failure = FALSE, output = FALSE)
}
\description{
Valor Limite anual = Valor Utilizado + Valor Transferido
}
