% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_limite_plurianual.R
\name{check_limite_plurianual}
\alias{check_limite_plurianual}
\title{Valor Limite anual = Valor Utilizado + Valor Transferido}
\usage{
check_limite_plurianual(
  base_limite_cota,
  stop_on_failure = FALSE,
  output = FALSE,
  json_outfile = NULL,
  log_level = "ERROR",
  msg_template = NULL
)
}
\description{
O valor do limite anutal não levará em consideração o valor transferido
quando se tratar dos anos 1, 2 e 3 do plurianual.
}
