% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_objetivos_estrategicos_exists.R
\name{check_objetivos_estrategicos_exists}
\alias{check_objetivos_estrategicos_exists}
\title{Verifica se cada programa possui pelo menos um objetivo estratégico}
\usage{
check_objetivos_estrategicos_exists(
  programas_planejamento,
  output = FALSE,
  stop_on_failure = FALSE,
  json_outfile = NULL,
  log_level = "ERROR",
  msg_template = NULL
)
}
\description{
Verifica se cada programa possui pelo menos um objetivo estratégico
}
