% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ods.R
\name{check_ods}
\alias{check_ods}
\title{Verificar vinculação dos programas aos Objetivos de Desenvolvimento Sustentável (ODS)}
\description{
São realizadas duas verificações:
\itemize{
\item cada programa possui pelo menos um ODS com exceção do programa padronizado 705;
\item se existem programas com ODS 18 - Não Possui ODS, conjugado com outro(s) ODS(s).
}
}
