% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ods_exists.R
\name{check_ods_exists}
\alias{check_ods_exists}
\title{Verifica cada programa possui pelo menos um ODS}
\usage{
check_ods_exists(
  programas_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Com exceção do programa 705 todos os demais
devem possuir um ou mais Objetivos de Desenvolvimento Sustentável (ODS)
}
