% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_projetos_estrategicos.R
\name{check_projetos_estrategicos}
\alias{check_projetos_estrategicos}
\title{Verificar se as ações estratégicas possuem algum projeto estratégico}
\usage{
check_projetos_estrategicos(
  acoes_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verificar se as ações estratégicas possuem algum projeto estratégico
}
