% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_limite.R, R/check_repasse_recursos.R
\name{check_repasse_recursos}
\alias{check_repasse_recursos}
\title{Valor Limite anual = Valor Utilizado + Valor Transferido}
\usage{
check_repasse_recursos(
  base_qdd_fiscal,
  base_repasse_recursos,
  stop_on_failure = FALSE,
  output = FALSE
)

check_repasse_recursos(
  base_qdd_fiscal,
  base_repasse_recursos,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Verificar se o valor total do orçamento (orçamento fiscal e orçamento de
investimento das empresas controladas) coincide com a projeção do PPAG
para o ano seguinte
}
