% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_setor_governo.R
\name{check_setor_governo}
\alias{check_setor_governo}
\title{Verificar se os setores de governo estão corretos}
\usage{
check_setor_governo(
  acoes_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verificar se os setores de governo estão corretos
}
