% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valores_qdd_plurianual.R
\name{check_valores_qdd_plurianual}
\alias{check_valores_qdd_plurianual}
\title{Total do orçamento fiscal e investimento SIGPLAN vs SISOR}
\usage{
check_valores_qdd_plurianual(
  base_qdd_plurianual,
  acoes_planejamento,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Verificar se o valor total do orçamento (orçamento fiscal e orçamento de
investimento das empresas controladas) coincide com a projeção do PPAG
para o ano seguinte
}
