% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valores_sigplan_programas.R
\name{check_valores_sigplan_programas}
\alias{check_valores_sigplan_programas}
\title{Verificar se os totais dos valores orçamentários e físicos de cada ano do PPAG estão consistentes entre as bases}
\usage{
check_valores_sigplan_programas(
  acoes_planejamento,
  programas_planejamento,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Verificar se os totais dos valores orçamentários e físicos de cada ano do PPAG são iguais
entre as bases "Programas", "Ações" e "Localizadores" com agrupamento pelas colunas:
\itemize{
\item orgão
\item unidade orçamentária
\item função
\item subfunção
\item programa
\item projeto, atividade ou operação especial
\item categoria econômica
\item grupo de despesa
\item modalidade de aplicação
\item elemento de despesa
\item identificador de ação governamental (IAG)
\item fonte de recurso
\item identificador de procedência e uso (IPU)
}
}
