% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valores_sisor.R
\name{check_valores_sisor}
\alias{check_valores_sisor}
\title{Consistência entre "QDD Fiscal" e "Orçamento da Despesa Fiscal / Itens de Despesa"}
\usage{
check_valores_sisor(
  base_qdd_fiscal,
  base_orcam_despesa_item_fiscal,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Verificar se o valor total do orçamento é igual entre as bases
"QDD Fiscal" e "Orçamento da Despesa Fiscal / Itens de Despesa" com
agrupamento pelas classificações:
\itemize{
\item orgão
\item unidade orçamentária
\item função
\item subfunção
\item programa
\item projeto, atividade ou operação especial
\item categoria econômica
\item grupo de despesa
\item modalidade de aplicação
\item elemento de despesa
\item identificador de ação governamental (IAG)
\item fonte de recurso
\item identificador de procedência e uso (IPU)
}
}
