% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_area_tematica_exists.R
\name{check_area_tematica_exists}
\alias{check_area_tematica_exists}
\alias{check_area_tematica_exists_programas}
\alias{check_area_tematica_exists_acoes}
\alias{check_area_tematica_exists_localizadores}
\title{Área temática por programa}
\usage{
check_area_tematica_exists_programas(
  programas_planejamento,
  stop_on_failure = FALSE,
  output = FALSE
)

check_area_tematica_exists_acoes(
  acoes_planejamento,
  stop_on_failure = FALSE,
  output = FALSE
)

check_area_tematica_exists_localizadores(
  localizadores_todos_planejamento,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Verifica se cada programa possui apenas uma área temática. É obrigatório a existência de uma área por programa (ie. \emph{one and only one}).

A verificação é realizada por base de dados:
\itemize{
\item programas_planejamento (\code{\link[=check_area_tematica_exists_programas]{check_area_tematica_exists_programas()}})
\item acoes_planejamento (\code{\link[=check_area_tematica_exists_acoes]{check_area_tematica_exists_acoes()}})
\item localizadores_todos_planejamento (\code{\link[=check_area_tematica_exists_localizadores]{check_area_tematica_exists_localizadores()}})
}
}
