% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_area_tematica_exists_localizadores.R
\name{check_area_tematica_exists_localizadores}
\alias{check_area_tematica_exists_localizadores}
\title{Verifica se cada programa possui apenas uma área temática - localizadores_todos_planejamento}
\usage{
check_area_tematica_exists_localizadores(
  localizadores_todos_planejamento,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
É obrigatório a existência de uma área por programa (ie. \emph{one and only one}).
}
