% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count_acoes.R
\name{check_count_acoes}
\alias{check_count_acoes}
\title{Verificar quantidade de ações entre as bases}
\usage{
check_count_acoes(
  acoes_planejamento,
  localizadores_todos_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verificar quantidade de ações entre as bases
}
