% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count_acoes_is_deleted.R
\name{check_count_acoes_is_deleted}
\alias{check_count_acoes_is_deleted}
\title{Verificar quantidade de ações excluídos entre as bases}
\usage{
check_count_acoes_is_deleted(
  acoes_planejamento,
  localizadores_todos_planejamento,
  output = FALSE,
  stop_on_failure = FALSE,
  json_outfile = NULL,
  log_level = "ERROR",
  msg_template = NULL
)
}
\description{
Se, no processo de exclusão de ações que foram incluídos e excluídos
durante o mesmo ciclo (momento R), o mesmo não tiver sido realizado no
SIGPLAN, esse teste vai apontar problemas.
}
