% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count_programas.R
\name{check_count_programas}
\alias{check_count_programas}
\title{Verificar quantidade de programas entre as bases}
\usage{
check_count_programas(
  programas_planejamento,
  acoes_planejamento,
  localizadores_todos_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verificar quantidade de programas entre as bases
}
