% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count_programas_is_deleted.R
\name{check_count_programas_is_deleted}
\alias{check_count_programas_is_deleted}
\title{Verificar quantidade de programas excluídos entre as bases}
\usage{
check_count_programas_is_deleted(
  programas_planejamento,
  acoes_planejamento,
  localizadores_todos_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Se o processo de exclusão de programas que foram incluídos e excluídos
durante o mesmo ciclo (momento R) não tiver sido realizado no SIGPLAN
esse teste vai apontar problemas.
}
