% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count_programas_is_new.R
\name{check_count_programas_is_new}
\alias{check_count_programas_is_new}
\title{Verificar quantidade de programas incluídos entre as bases}
\usage{
check_count_programas_is_new(
  programas_planejamento,
  acoes_planejamento,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
Verificar quantidade de programas incluídos entre as bases
}
