% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/check_detalhamento_obras_numero_siad_duplicated.R
\name{check_detalhamento_obras_numero_siad_duplicated}
\alias{check_detalhamento_obras_numero_siad_duplicated}
\title{Verifica se número da obra siad possui valores duplicados por ação}
\usage{
check_detalhamento_obras_numero_siad_duplicated(
  base_detalhamento_obras,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Verifica se número da obra siad possui valores duplicados por ação
}
