% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fechamento_fonte_orcam_fiscal.R
\name{check_fechamento_fonte_orcam_fiscal}
\alias{check_fechamento_fonte_orcam_fiscal}
\title{Fechamento receita e despesa por fonte do orçamento fiscal}
\usage{
check_fechamento_fonte_orcam_fiscal(
  base_orcam_receita_fiscal,
  base_qdd_fiscal,
  stop_on_failure = FALSE,
  output = FALSE
)
}
\description{
Fechamento receita e despesa por fonte do orçamento fiscal
}
