% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_intra_orcamentaria.R
\name{check_intra_orcamentaria}
\alias{check_intra_orcamentaria}
\title{Receitas e Despesas Intra-Orcamentárias}
\description{
Verificar se as operações intra-orçamentárias registradas na
Base Intraorçamentária de Repasse estão adequadamente detalhadas no orçamento
}
