% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_setor_governo.R
\name{check_setor_governo}
\alias{check_setor_governo}
\title{Verificar se os setores de governo estão corretos}
\usage{
check_setor_governo(
  acoes_planejamento,
  aux_setor_governo,
  output = FALSE,
  stop_on_failure = FALSE
)
}
\description{
A referência correta dos setores de governo é disponibilizada
pela DCPPN via tabela auxiliar no recurso
\href{https://github.com/splor-mg/volumes-loa-dados}{\code{volumes-loa-dados.desc_setor_governo}}
}
